reverted. double to float

+++ libraries/AP_Math/vector3.h
@@ -253,5 +253,5 @@ typedef Vector3<int16_t>                Vector3i;
 typedef Vector3<uint16_t>               Vector3ui;
 typedef Vector3<int32_t>                Vector3l;
 typedef Vector3<uint32_t>               Vector3ul;
-typedef Vector3<double>                  Vector3f;
+typedef Vector3<float>                  Vector3f;
 typedef Vector3<double>                 Vector3d;
